+++ frontend/src/lib/docs.js
@@ -1,0 +1,2 @@
+export const REPO = 'https://github.com/jasonhaymond/haven-backup'
+export const DOCS = `${REPO}/blob/master/docs`